msm: krait-regulator: add control for ldo/hs for krait
There is a single voltage supply line from the pmic for all the kraits
and within each krait susbsystem there is a LDO and a block head switch
(BHS) in parallel which control the voltage to that krait. In general,
for maximum efficiency when the voltage requirement of the core is low,
LDO should be used. BHS switch should be used when the voltage
requirement is high or the voltage supply from the pmic line is not
high enough for operating the LDO.

Add code to control the LDO or BHS configuration for krait regulators.

Change-Id: Ia1163f505f2cea57d29dc11fed71853a47be2aab
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>