msm: SSR: Make subsystem restart into a bus
With a bus type in place for subsystem restart we can easily
discern what subsystems are present and see what state they are
in (either online or offline at the moment). We also support
polling of the state sysfs node so that programs can monitor
state transitions.

This is mostly a copy paste job from PIL with a 's/pil/subsys/g'.

Change-Id: I93443bbc9748468d6fc51d96ce4a3872617cec36
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>