power: pm8921-charger: enable battery alarm
During UVLO events the charger hardware may lock up
if it is in hardware clock switching mode.

To fix this add a battery alarm with a lower threshold
that disables hardware clock switching.

When the system is awake the resume callback ensures
that the charger is running off of the 19.2 MhZ clock
which does not allow the charger to lock up. Therefore
if the battery alarm wakes up the device before hitting
UVLO the charger hardware will not be in hardware clock
switching mode.

(cherry picked from commit 8b079fb38a52cc0bc435d57b21758cf6b2dd5caf)

Change-Id: I0c9c5ad01b5bdbb4ec47b27c4b31d7ccb6129b1a
CRs-Fixed: 374607
Signed-off-by: David Keitel <dkeitel@codeaurora.org>