msm: subsystem_restart: Don't use phase 3 SSR on the SGLTE
On the SGLTE platform, phase 2 SSR is required to address
the dependency between the internal and external modems. The
other subsystems (RIVA, LPASS etc) still need to do an
independent restart. In the absence of a matching restart
order, a request for a subsystem restart on a subsystem will
be treated as an  independent phase-3 like restart, even if
the current restart_level is phase 2. Therefore, don't allow
phase 3 on SGLTE platforms.

Change-Id: Idbae84bd6624390adb95a80ddfb0f963d8ee647d
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>