msm: krait-regulator: disable phase control of the ganged regulator
Due to a bug in the FTS2 (fast transient switch version 2) digital,
the bus used for collecting phase balancing data is not properly getting
reset on phase count transitions.

Until this is fixed in the later pmic hardware versions, delete the
code that changes the phases. The bootloader will turn on all the phases
and the krait-regulator will only change the voltage on the pmic gang.

Change-Id: I7a88a7afdcad3b722a10b080d6d1e2c55723c5fc
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>