msm: clock-8974: Make mmss_s0_axi_clk depend on mmssnoc_axi_clk
The mmss_s0_axi clock needs to be turned on to allow MMSS NOC
slaves to be able to access memory across the MMSS NOC and
System NOC interconnect. However, this clock needs the
mmssnoc_axi_clk to be on for the NOC handshake to work.

Therefore, have the bus driver turn on the mmss_s0_axi_clk when
there are bandwidth requests for the MMSS NOC AXI clock and have
mmss_s0_axi_clk depend on the mmssnoc_axi_clk.

Change-Id: I81d0dc27ea7a7983ddbebbfff67c3426b3c6da94
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>