msm: pil-q6v5: Enable q6ss_ahbm_clk for LPASS Q6
This clock must be enabled for the LPASS Q6 to reach its own clock
registers, so that it can turn on additional clocks on its own. Do
this.

To avoid adding another clock to struct q6v5_data, rename the mem_clk
member in it (used by MSS) to something more generic so it can also
be used by LPASS.

Change-Id: I0b4ec626d249a48dd832a818f630f1f6aed1f98a
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>