msm: 8974: Add power support for KS8851 SPI Ethernet device
Add a fixed-regulator device tree node to support the KS8551 SPI
Ethernet peripheral.

Configure pm8941 MPP5 for digital output active-high, since it is
connected to the SHDN_N pin on the MAX1818. In order to enable
the regulator, it's necessary to supply the SHDN_N pin with
greater than 1.6V. Since pm8941_s3 is at 1.8V and always enabled,
use this as the VIN supply.

Change-Id: I9166041a3a12fc8e789866ae1d972047996e8027
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>