msm: krait-regulator: Change default min voltage to 0.9V
The current default min voltage value of 0.5V is inconsistent with
the PMIC default for 8841_s5 of 0.9V. This discrepancy causes
krait-regulator to incorrectly set kraits to LDO mode instead of
headswitch, despite the default hardware voltage being too high to
support LDO mode.

The bootstrap procedure is guaranteed to never to change 8841_s5
to a different voltage, so it's safe and reasonable to assume a
default value instead of reading 8841_s5's real-time value on
probe().

Change-Id: I28a222574d47c4961ebc2dee9c9f40d687cc4d55
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>