msm: msm_bus: Fix scaling of ocmem core clock
An optimization in bus driver prevents checking of slave clocks if
the clock of the NoC on which the slave is connected is not
touched.
Since ocmem slave is represented on a virtual NoC, the ocmem core
clock was not being scaled. Fix this by checking for status of
slave clocks.

Change-Id: I42af40b9f414a62c3712850340eef203ad6ad33d
Signed-off-by: Gagan Mac <gmac@codeaurora.org>