adiag: Add new entry in message mask table
DIAG driver maintains a table of F3 masks. This table
has to be in sync with modem mask information. Modem
updates are not synchronized with Linux code updates.
Hence, sometimes modem and Linux go out of sync. This
results in missing F3 messages.

Adding this entry will resolve the dependency of manual
syncing and will make mask updates dynamic.

Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>

Change-Id: I37f163aa1560048249da8aa66842a3852bb81ddd
Signed-off-by: Neha Pandey <nehap@codeaurora.org>