ASoC: WCD9310: Defer the setup of headphone removal detection
After the inserted accessory is detected as a valid headphone,
polling is performed on the microphone to correct the accessory
detection in case it was wrongly identified as a headphone.

It was observed that at the end of polling, headphone removal
interrupt was fired, causing headphone playback to stop. Fixed
by deferring the headphone removal setup until either polling is
complete or polling detects a removal

CRs-fixed: 384967
Change-Id: I558371beb0b0206f3857f4c45967cbb60fe55b03
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>