msm: rpm-smd: Remove BUG if packet size is 0
When a interrupt for data event is received, it is possible that the SMD
packet could have been processed on other core executing with interrupts
locked. In this case, the data event would have scheduled a workqueue
which would result in 0 bytes for packet size. This is a valid scenario
and shouldn't throw a bug.

Change-Id: I8cdb411491e1b344356bae7071ec65342e456655
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>