msm: camera: Update the missing information for mctl pp node buffers.
When a buffer is dequeued from mctl pp node, the offset information is
not copied over to the userspace. This will result in garbage offsets
received by the user when he tries to dequeue a buffer on the mctl pp
node. Update these information when the buffer is being dequeued.

Change-Id: I93212ce784e5cc2cb062f57873c28fb2e0cf6ed5
Signed-off-by: Kiran Kumar H N <hurlisal@codeaurora.org>