gpio: qpnp-pin: Add support for gpio major revision 1
In revision 1 of the gpio block, the GPIO_STATUS1 GPIO_OK changes
from bit 1 to bit 7. Update the code to reflect this.

Also verify that PMIC pins have valid types and subtypes upon
probe. This allows us to make assumptions about correctness later
on when checking for capabilities of a pin in question.

Change-Id: I7c8e4c5e44e00ff1c59ac8126449a1461e76f6d8
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>