power: pm8921-charger: workaround for active charger path
The active path bits in PBL_ACCESS1 register don't reflect the actual
status until some delay. This delay varies from few millisecond to
three seconds.

The unplug detection loop relies on these bits reflecting the correct
status for proper operation. Due to this bug the unplug worker wont
run when a charger is plugged in.

Update the code such that if it sees nothing set in the active path,
it still reschedules the unplug detection worker.

Change-Id: Ib99cd49623698d0295bc3c3cf880225495d530cb
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>