usb: ks_bridge: Check for received data upon unlinking the urb
There is a possibility of runtime suspend in the middle of
data transfer on the rx data path of driver. This results in
the completion of data transfer on bulk IN endpoint with short
packet or integral multiple of 512 bytes. Hence for -ENOENT
transfer status check for non-zero data length and pass it to
the consumer in the userspace. Also, replace printk with
DBG-level message in driver suspend function and in rx
callback function for -EPROTO urb status.

CRs-Fixed: 394954
Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>
(cherry picked from commit 853ebf649fbd19362c22a23e3c14ffa6cc2d7c5f)

Conflicts:

	drivers/usb/misc/ks_bridge.c

Change-Id: I1b976d1bbd9acd09b46b51d433a41f904001c746
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>