msm: cpr: Clear the auto NACK down interrupt bit in freq. transition
CPR driver enables the auto NACK down interrupt bit right now
when the CPR suggests the voltage less than the Vmin limit
specified by the driver. It only gets re-enabled if we receive
the one UP interrupt. This may not work in the case when
the freq. transition happens where the DCVS driver sets
the new voltage and CPR suggests the correct down voltage
recommendation, so clear the NACK down enable bit during
the post change phase of the freq. transition.

Change-Id: Iee6cc12b60acdc3dec88d147b39262befcb8ce0d
Signed-off-by: Trilok Soni <tsoni@codeaurora.org>