msm: audio: qdsp6v2: Fix WMV stream resume issue
- Suspend phone while WMV clip playback is going on.
Resume it, clip fails to play.
- At resume of WMV clip, Pause is issued initially
to driver. But there is a bug in pause API which is
not returning appropriate result. OMX component when
starts playback, it now issues resume to driver and
driver is setting DSP to run state even before ENC_DEC
params are set.
- Fix pause code to return error value if driver is not
enabled. This will avoid driver in sending CMD_RUN
before ENC_DEC command.

CRs-Fixed: 394756
Change-Id: Ic5f0b4fc92b2892885fe2ff2c86593bd269041ea
Signed-off-by: Laxminath Kasam <lkasam@codeaurora.org>