power_supply: add BMS for a new power supply type
Add POWER_SUPPLY_TYPE_BMS for the hardware chip or module which
monitors battery for state of charge and other properties.

BMS will register itself as a separate power supply of type BMS.
The charger will register the battery power supply and pull BMS data
from the BMS power supply if it exists.

Change-Id: I071f0d958a4b2b6af5f51460d201f6970aed595f
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>