msm: dcvs: add thermal control
Add capability to read the sensor data and report the temperature of
a core to the TZ.

Note that this was done earlier whenever a frequency update was done,
instead create a delayed workqueue per core. Also create a attribute
that controls the rate of thermal update per core.

If that attribute is not set use 60seconds as the default update rate.
If it is less than one second, default to using one second.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit 80add4203ec4538046ca926e2747db6d01c14783)

Change-Id: If96927dc48e0d029d0148eccda34366b470f0437
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>