ASoC: WCD9304: Wait for earpiece PA to settle down
During playback of audio on earpiece path, the PA on the path
is the last component to be turned on. The turn on of PA can
cause a glitch in the output signal because the level of the PA
is not yet settled. Waiting for 20 msec after turning on the PA
avoids this glitch.

Change-Id: I7fb124893d240afe92a2d4af403b9b847a9272b9
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
(cherry picked from commit 815090d6eee6662e24b0d58b7d8c2d8faa16b413)