sysmon: check if mutex lock is available
Modem boots up completedly and opens the logical channel.  Mutex is initialized
at that time.  If sysmon_get_reason is accessed before that, panic occurs by
accessing the uninitialized mutex.  Since this is a valid access scenario, we
add a flag to check if the subsystem has been probed, and return an error if
sysmon_get_reason() gets called before that.

Change-Id: If81f9fc1be5b9e0e9e6e69366d1b8d6df35e2857
Signed-off-by: Iliyan Malchev <malchev@google.com>