qup_i2c: Add runtime PM support to I2C QUP controller
Runtime PM is used to turn clocks on and off when the
controller is not in use. When the controller is used
for messaging, pm_runtime_get_sync is called before
sending the message. pm_runtime_mark_last_busy and
pm_runtime_put_autosuspend are called when message
is sent and/or response is received.

(cherry picked from commit ebe211daa18c5bd65d1d91ae8fdc40a7ffee506f)

Change-Id: I4cc8dabbdf992b890bdc6a1699f508d76777e9ff
CRs-Fixed: 406910
Signed-off-by: Alok Chauhan <alokc@codeaurora.org>
Signed-off-by: Manish Kumar <manishku@codeaurora.org>