qseecom: Check if app exists before loading the app
There is a chance of send a request to QSEE to load an app that
is already previously loaded. Requesting QSEE to load an app
that is already loaded will result in a failure.

Issuing a request to QSEE to check if app already exists before
issuing a request to load it, all within the same app_access_lock
mutex, fixes this issue.

Change-Id: Id32b34fec654f82cf0247520d4f2a56801b4c385
CRs-Fixed: 420473
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>