mmc: use percentage changed sectors as indication to check BKOPS need
In order to keep card endurance, it is best not to check the need for
BKOPS every time the MMC is idle.
However, the decision when to check the card need for BKOPS according
to number of changed sectors doesn't fit all cards sizes. Therefore,
the check for BKOPS need is based on percentages of changed sectors
of the card size.
In order to assure that BKOPS will be triggered again in case it was
interrupted, the accumulated number of changed sectors is cleared only
in case BKOPS was completed without interruption.

Signed-off-by: Maya Erez <merez@codeaurora.org>
(cherry picked from commit 7701a8fe14048b8eb0e7ff1cd33eb8c915c14e17)

Change-Id: Ibd3066d63fbdfd0c3cc9af9c2daf28183b20ca86
Signed-off-by: Maya Erez <merez@codeaurora.org>