power: pm8921-charger: force 19.2Mhz while writing to a register
We saw a problem where the charger was locked up and it reported
incorrect FSM states.

The root cause was that PMIC could be doing short sleeps while
LPM is set to 0. This means it would enter 32Khz and exit 32Khz
fairly quickly causing the lockup.

For e.g.
pm_chg_write gets called
write LPM = 0
pmic enters and exits sleep quickly.
charger locks up
the writes and reads are unreliable from this point on.

To prevent the lock up instead of using the LPM bit force
the charger to run from 19.2Mhz while doing a write.

Change-Id: I7e61abc394f91eb0c4c83ea1f762972ed1c0ff78
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Anil Gahlot <anilg@codeaurora.org>