msm_fb: hdmi: Do not turn on HDMI if not ready
At bootup, if HPD is not configured, HDMI is not in
the state of getting turned on. This change makes sure
HDMI is not turned on/off unnecessarily which turns on
clocks as well.

CRs-Fixed: 447511
Change-Id: I074d9c463943770775b060c4a93b9120dfe9bcf0
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
(cherry picked from commit fb1b974d191cea51d1377c2917c8f0cb36d0311e)

Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>