power: pm8921-charger: don't reject all the chg current requests
The driver currently rejects any changes to the USB charge current if
no other source of power is present. This causes issues because if a
pc usb cable which supplies 500mA  is connected while the PMIC was
drawing 1500mA, it will continue to draw 1500mA from it causing charger
collapse.

The reason why the rejects were put in place were because if we are
running from USB as the source, vbus_draw() could be called with 0mA.
This would cause brownouts.

So only reject setting the current to 0mA in case another source is not
present.

Change-Id: I674d9d17b329c72de44cfce979a312d180529ac8
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Anil Gahlot <anilg@codeaurora.org>