power: pm8921-bms: report higher soc even when called within a second
The current code puts a restriction in the charger scaling algorithm
that state of charge (soc) should be called after a second. If it is
called earlier, simply report the previous soc.

There are user space applications that call soc every few milliseconds.
The soc appears stuck for such applications.

Remove that restriction.

Change-Id: I22d2dbdae5c9c859d4baca2bdb7bea71da09b85b
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>