power: pm8921-bms: update power supply class if soc changes
There is a hearbeat loop that queries the state of charge
(soc) every few seconds. If the soc changes multiple times in
that period, we miss updating the userspace of those changes.

The userspace will notice a big step in soc, which is undesirable.
Update the userspace when the calculated SOC changes.

Change-Id: I2c149e6edd7c194e7f31ece95e2003ffde935806
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>